Started Day 12
Still working on a solution to Day 12. This is by far the hardest challenge yet. So far I can create a map of the input data using a dictionary. I can also create a nest list of paths. I think next is to create a function to unnest the list to clearly display the paths (mostlyy for troubleshooting). I also have to figure out how to return to a big letter when generating the paths.